Folder restructuring and removed unused screens
