+++ src/screens/Home/components/CurrencyInfo.tsx
