+++ src/screens/Home/components/HeaderDetails/components/CurrentPlanView.tsx
