+++ src/screens/Home/components/HeaderDetails/components/HeaderBar.tsx
