+++ src/screens/Home/components/HeaderDetails/components/UAIView.tsx
