--- src/screens/HomeScreen/components/HomeScreenWrapper.tsx
@@ -1,26 +1,0 @@
-import { StyleSheet } from 'react-native';
-import React from 'react';
-import { Box, StatusBar, useColorMode } from 'native-base';
-import HeaderDetails from './HeaderDetails';
-
-function HomeScreenWrapper({ children }) {
-  const { colorMode } = useColorMode();
-  return (
-    <Box
-      style={[styles.container, { backgroundColor: colorMode === 'light' ? '#F2EDE6' : '#323C3A' }]}
-    >
-      <StatusBar barStyle={colorMode === 'light' ? 'dark-content' : 'light-content'} />
-      <HeaderDetails />
-      <Box style={styles.container}>{children}</Box>
-    </Box>
-  );
-}
-
-export default HomeScreenWrapper;
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    paddingHorizontal: 10,
-  },
-});
